+++ yash-builtin/src/wait/core.rs
@@ -22,8 +22,8 @@
 
 use thiserror::Error;
 use yash_env::job::Pid;
+use yash_env::signal;
 use yash_env::system::Errno;
-use yash_env::trap::Signal;
 use yash_env::Env;
 use yash_env::System as _;
 use yash_semantics::trap::run_trap_if_caught;
@@ -36,8 +36,8 @@ pub enum Error {
     NothingToWait,
     /// The built-in was interrupted by a signal and the trap action was
     /// executed.
-    #[error("trapped({0})")]
-    Trapped(Signal, yash_env::semantics::Result),
+    #[error("trapped signal {0}")]
+    Trapped(signal::Number, yash_env::semantics::Result),
     /// An unexpected error occurred in the underlying system.
     #[error("system error: {0}")]
     SystemError(#[from] Errno),
@@ -195,7 +195,7 @@ mod tests {
 
                 // Now the function should return.
                 let result = future.await;
-                assert_eq!(result, Err(Error::Trapped(Signal::SIGTERM, Continue(()))));
+                assert_eq!(result, Err(Error::Trapped(SIGTERM, Continue(()))));
             }
 
             // The trap action must have assigned the variable.