+++ yash-builtin/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to `yash-builtin` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.1] - Unreleased
+## [0.2.0] - Unreleased
 
 ### Added
 
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `trap::display_traps` is now marked `#[must_use]`.
 - `trap::display_traps` is now additionally takes a
   `yash_env::trap::SignalSystem` argument.
+- `wait::core::Error::Trapped` now contains a `yash_env::signal::Number`
+  instead of a `yash_env::trap::Signal`.
 
 ## [0.1.0] - 2024-04-13